backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "django-ninja>=1.4.3",
    "django-unfold>=0.64.2",
    "environs>=14.3.0",
    "gunicorn>=23.0.0",
    "httptools>=0.6.4",
    "loguru>=0.7.3",
    "psycopg>=3.2.9",
    "psycopg-binary>=3.2.9",
    "uvicorn>=0.35.0",
    "uvicorn-worker>=0.3.0",
    "uvloop>=0.21.0",
]

[dependency-groups]
dev = [
    "django-stubs>=5.2.2",
]

# ====================
# Ruff configuration
# ====================

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = [".venv", "__pycache__", ".git", ".mypy_cache"]

[tool.ruff.lint.isort]
known-first-party = ["app", "contexts", "core", "lib", "providers"]
combine-as-imports = true
force-sort-within-sections = true
lines-after-imports = 2

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "lf"
